Bumped the springboot version from 3.1.12 to 3.4.5

+++ src/main/java/com/divjazz/recommendic/user/service/ConsultantService.java
@@ -127,6 +127,9 @@ public class ConsultantService {
     }
 
     public void deleteConsultantById(String userId) {
+        if (userService.isUserNotExists(userId)) {
+            throw new EntityNotFoundException("Consultant does not exist or has been deleted");
+        }
         consultantRepository.deleteByUserId(userId);
     }
